# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.PaginatedCategories:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Category'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UpdateMediaInput:
    properties:
      altText:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: 'REST API for Product & Category Service. Management routes require
    an Auth0 RS256 bearer access token whose exact permission matches the route; denials
    use the {"error":{"code":"...","message":"...","details":{}}} envelope with 401
    UNAUTHENTICATED for a missing, malformed, invalid, expired, or wrong-audience
    credential and 403 FORBIDDEN for a valid token without the route''s exact permission.
    Public exception: GET /categories, GET /categories/{id}, GET /categories/dropdown,
    GET /products, and GET /products/{id} require no token and ignore any supplied
    Authorization header. Customer exception: POST /products/{id}/reviews, PATCH /reviews/{reviewId},
    DELETE /reviews/{reviewId}, and GET /users/me/reviews keep the existing customer
    HS256 token and ownership rules, and an Auth0 permission is not a substitute for
    customer identity.'
  title: Gin Product Service API
  version: "1.0"
paths:
  /categories:
    get:
      description: Get a paginated, filterable list of all existing categories
      parameters:
      - description: Filter by partial (case-insensitive) name match
        in: query
        name: name
        type: string
      - description: Page number (1-indexed)
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: 'Sort column. One of: name, created_at'
        in: query
        name: sort_by
        type: string
      - description: 'Sort direction. One of: asc, desc'
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PaginatedCategories'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Category to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Soft-delete an existing product category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a category (soft delete)
      tags:
      - categories
    get:
      description: Get a single category by its numeric ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing product category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
  /categories/dropdown:
    get:
      description: Get a lightweight list (id + name) of categories for dropdown UI
      parameters:
      - description: Filter by partial (case-insensitive) name match
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CategoryOption'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List categories for dropdown
      tags:
      - categories
  /inventory/reservations:
    post:
      consumes:
      - application/json
      description: Reserve sufficient stock for one order at the default fulfillment
        location. Idempotent by orderId; the order service supplies the hold expiry.
      parameters:
      - description: Checkout reservation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReservationSuccessResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReservationSuccessResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an atomic checkout inventory reservation
      tags:
      - inventory
  /products:
    get:
      description: Get a paginated, filterable list of products
      parameters:
      - description: Filter by partial (case-insensitive) title, handle or category
          name match
        in: query
        name: search
        type: string
      - description: Filter by category ID (0 means no filter)
        in: query
        name: category_id
        type: integer
      - description: 'Filter by status. One of: draft, active, archived'
        in: query
        name: status
        type: string
      - description: Minimum variant price (inclusive). Products must have a variant
          priced >= minPrice
        in: query
        name: minPrice
        type: number
      - description: Maximum variant price (inclusive). Products must have a variant
          priced <= maxPrice
        in: query
        name: maxPrice
        type: number
      - description: Page number (1-indexed)
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: 'Sort column. One of: title, created_at, category_name, price,
          popularity'
        in: query
        name: sort_by
        type: string
      - description: 'Sort direction. One of: asc, desc'
        in: query
        name: sort_dir
        type: string
      - description: Comma-separated list of rounded average ratings to filter by,
          e.g. rating=3,4,5. Each value must be 1-5.
        in: query
        name: rating
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PaginatedProducts'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a product with options, variants, gallery and initial stock
        in one transaction
      parameters:
      - description: Product to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a product
      tags:
      - products
  /products/{id}:
    delete:
      description: Soft-delete a product by setting its status to archived
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Archive a product (soft delete)
      tags:
      - products
    get:
      description: Get a single product (with options, variants and media) by UUID
        or handle
      parameters:
      - description: Product UUID or handle
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a product by ID or handle
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Update product header fields only (title, description, vendor,
        handle, categoryId, status). Omitted fields are left unchanged.
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - products
  /products/{id}/media:
    post:
      consumes:
      - application/json
      description: Upload/attach new media to a product's gallery (accepts an array)
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: Media to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BulkCreateMediaInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Attach media to a product gallery
      tags:
      - products
  /products/{id}/media/{media_id}:
    delete:
      description: Remove media from a product entirely (cascades variant media links)
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: Media UUID
        in: path
        name: media_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete product media
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Update a media item's metadata (alt_text)
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: Media UUID
        in: path
        name: media_id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMediaInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update media metadata
      tags:
      - products
  /products/{id}/media/reorder:
    patch:
      consumes:
      - application/json
      description: Bulk-update the positions of a product's gallery media
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: New positions
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReorderMediaInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reorder product media
      tags:
      - products
  /products/{id}/options:
    post:
      consumes:
      - application/json
      description: Create a new option with its initial values for a product
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: Option to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOptionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a product option
      tags:
      - products
...
  /products/{id}/options/{option_id}/values:
    post:
      consumes:
      - application/json
      description: Add a new value to an existing option
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      - description: Option UUID
        in: path
        name: option_id
        required: true
        type: string
      - description: Value to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOptionValueInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add an option value
      tags:
      - products
...
  /users/me/reviews:
    get:
      description: Get a paginated list of the authenticated user's reviews. Requires
        authentication.
      parameters:
      - description: Page number (1-indexed)
        in: query
        name: page
        type: integer
      - description: Page size (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserReviewListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List the current user's reviews
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: Auth0 RS256 access token as `Bearer <token>`. The token must carry
      the operation's exact permission and be signed by a key currently published
      by the configured Auth0 domain. Token issuer and audience must match the configured
      Auth0 API; an invalid, expired, or unintended token yields 401.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"